requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
xgboost==3.1.1
gtfs-realtime-bindings==1.0.0
protobuf==6.33.0
requests==2.32.5
python-dotenv==1.0.0
pytz==2025.2
